services/gpt-context/pyproject.toml
[tool.poetry]
name = "gpt-context"
version = "0.1.0"
description = ""
authors = ["Pavel Karpovich <pavel.karpovich@outlook.com>"]
readme = "README.md"
packages = [{include = "gpt_context"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
tiktoken = "^0.4.0"
grpcio = "^1.54.2"
grpcio-tools = "^1.54.2"
betterproto = {version = "^2.0.0b5", extras = ["compiler"]}
redis = "^4.5.5"
langchain = "^0.0.228"
weaviate-client = "^3.22.0"
openai = "^0.27.8"
google-api-python-client = "^2.91.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
pymongo = "^4.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
